+++ changelog.md
@@ -5,6 +5,7 @@
 ### 2.0.0-alpha.13
 
 - Move back to the original URLs for diffs, instead of relying on PR metadata - orta
+- Updates the types for `schedule` to be more accepting of what it actually takes - orta
 
 ### 2.0.0-alpha.12